Dataset
This dataset encompasses a comprehensive record of Summer Olympic Games from the inaugural 1896 Athens Olympics to the most recent 2024 Paris Olympics. It provides a rich source of information about athletes, their performances, and the medals awarded over a span of more than a century.

This dataset is inspired by the Olympic Games' rich history and the evolution of global sports competition. By merging historical data with the latest records, this dataset allows for a comprehensive analysis of trends in Olympic performance, country-specific achievements, and the progression of various sports disciplines over time.